Moved OrdersWidget to apps/gp-v1/components
