+++ src/apps/gp-v1/components/OrdersWidget/OrderRow.styled.ts
