+++ src/apps/gp-v1/components/OrdersWidget/OrdersWidget.styled.ts
