+++ src/apps/gp-v1/components/OrdersWidget/useDeleteOrders.tsx
