+++ src/apps/gp-v1/components/TradeWidget/TradeWidget.styled.ts
@@ -1,7 +1,7 @@
 import styled from 'styled-components'
 import Widget from 'components/layout/LegacyTradeLayout/Widget'
 import { Wrapper } from 'components/ConnectWalletBanner'
-import { OrdersWrapper } from 'components/OrdersWidget/OrdersWidget.styled'
+import { OrdersWrapper } from 'apps/gp-v1/components/OrdersWidget/OrdersWidget.styled'
 import { WalletDrawerInnerWrapper } from 'apps/gp-v1/components/DepositWidget/Form.styled'
 
 import { MEDIA } from 'const'
