+++ src/apps/gp-v1/pages/Orders.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import OrdersWidget from 'components/OrdersWidget'
+import OrdersWidget from 'apps/gp-v1/components/OrdersWidget'
 import { PageWrapper } from 'components/layout'
 
 const Orders: React.FC = () => (
